+++ pkg/grpc/service/study-management_endpoints.go
@@ -4,6 +4,7 @@ import (
 	"context"
 	"fmt"
 	"reflect"
+	"sort"
 	"time"
 
 	"github.com/coneno/logger"
@@ -46,7 +47,7 @@ func (s *studyServiceServer) CreateNewStudy(ctx context.Context, req *api.NewStu
 	s.SaveLogEvent(req.Token.InstanceId, req.Token.Id, loggingAPI.LogEventType_LOG, constants.LOG_EVENT_STUDY_CREATION, req.Study.Key)
 
 	// create required indexes for the new study:
-	err = s.studyDBservice.CreateSurveyDefintionIndexForStudy(req.Token.InstanceId, study.Key)
+	err = s.studyDBservice.CreateIndexModelForSurveysForStudy(req.Token.InstanceId, study.Key)
 	if err != nil {
 		logger.Error.Printf("unexpected error when creating survey definition indexes: %v", err)
 	}
@@ -674,6 +675,151 @@ func (s *studyServiceServer) UpdateResearcherNotificationSubscriptions(ctx conte
 	return subscriptions, nil
 }
 
+func (s *studyServiceServer) RunRulesForPreviousResponses(ctx context.Context, req *api.RunRulesForPreviousResponsesReq) (*api.RuleRunSummary, error) {
+	if req == nil || token_checks.IsTokenEmpty(req.Token) || req.StudyKey == "" {
+		return nil, status.Error(codes.InvalidArgument, "missing argument")
+	}
+
+	if !token_checks.CheckRoleInToken(req.Token, constants.USER_ROLE_ADMIN) {
+		err := s.HasRoleInStudy(req.Token.InstanceId, req.StudyKey, req.Token.Id,
+			[]string{types.STUDY_ROLE_MAINTAINER, types.STUDY_ROLE_OWNER},
+		)
+		if err != nil {
+			s.SaveLogEvent(req.Token.InstanceId, req.Token.Id, loggingAPI.LogEventType_SECURITY, constants.LOG_EVENT_RUN_CUSTOM_RULES, fmt.Sprintf("permission denied for running custom rules in study %s  ", req.StudyKey))
+			return nil, status.Error(codes.Internal, err.Error())
+		}
+	}
+
+	start := time.Now().Unix()
+
+	type Counters struct {
+		Participants                  int32
+		ParticipantStateChangePerRule []int32
+	}
+	counters := &Counters{
+		Participants:                  0,
+		ParticipantStateChangePerRule: make([]int32, len(req.Rules)),
+	}
+
+	// Convert rules from API type:
+	rules := make([]*types.Expression, len(req.Rules))
+	for index, rule := range req.Rules {
+		rules[index] = types.ExpressionFromAPI(rule)
+	}
+
+	err := s.studyDBservice.FindAndExecuteOnFilteredParticipantsStates(
+		ctx,
+		req.Token.InstanceId,
+		req.StudyKey,
+		req.Filter.ParticipantIds,
+		req.Filter.ParticipantStatus,
+		func(dbService *studydb.StudyDBService, p types.ParticipantState, instanceID, studyKey string, args ...interface{}) error {
+
+			counters.Participants += 1
+
+			participantID2, _, err := s.profileIDToParticipantID(instanceID, studyKey, p.ParticipantID, true)
+			if err != nil {
+				logger.Error.Printf("RunRulesForPreviousResponses: %v", err)
+				return status.Error(codes.Internal, err.Error())
+			}
+
+			actionData := studyengine.ActionData{
+				PState:          p,
+				ReportsToCreate: map[string]types.Report{},
+			}
+
+			surveyKeys := req.Filter.SurveyKeys
+			if len(req.Filter.SurveyKeys) == 0 {
+				surveyKeys, err = s.studyDBservice.GetSurveyKeysInStudy(instanceID, studyKey, false)
+				if err != nil {
+					logger.Error.Printf("RunRulesForPreviousResponses: %v", err)
+					return status.Error(codes.Internal, err.Error())
+				}
+			}
+
+			anyChange := false
+			responses := []types.SurveyResponse{}
+
+			for _, surveyKey := range surveyKeys {
+				responses_temp, err := s.studyDBservice.FindSurveyResponses(instanceID, studyKey, studydb.ResponseQuery{
+					ParticipantID: p.ParticipantID,
+					SurveyKey:     surveyKey,
+					Since:         req.Filter.From,
+					Until:         req.Filter.Until,
+				})
+				if err != nil {
+					logger.Error.Printf("RunRulesForPreviousResponses: %v", err)
+					return status.Error(codes.Internal, err.Error())
+				}
+				responses = append(responses, responses_temp...)
+			}
+
+			//sort responses descending order
+			sort.SliceStable(responses, func(i, j int) bool {
+				return responses[i].SubmittedAt > responses[j].SubmittedAt
+			})
+
+			for i := len(responses) - 1; i >= 0; i-- {
+				for index, rule := range rules {
+					if rule == nil {
+						continue
+					}
+					event := types.StudyEvent{
+						Type:                                  "SUBMIT",
+						Response:                              responses[i],
+						InstanceID:                            instanceID,
+						StudyKey:                              studyKey,
+						ParticipantIDForConfidentialResponses: participantID2,
+					}
+
+					newState, err := studyengine.ActionEval(*rule, actionData, event, studyengine.ActionConfigs{
+						DBService:              s.studyDBservice,
+						ExternalServiceConfigs: s.studyEngineExternalServices,
+					})
+					if err != nil {
+						return err
+					}
+
+					if !reflect.DeepEqual(newState.PState, actionData.PState) {
+						counters.ParticipantStateChangePerRule[index] += 1
+						anyChange = true
+					}
+					actionData = newState
+				}
+				for key, report := range actionData.ReportsToCreate {
+					report.Timestamp = responses[i].SubmittedAt
+					actionData.ReportsToCreate[key] = report
+				}
+
+				s.saveReports(instanceID, req.StudyKey, actionData.ReportsToCreate, "")
+				actionData.ReportsToCreate = map[string]types.Report{} //clean up ReportsToCreate
+			}
+
+			if anyChange {
+				// save state back to DB
+				_, err := s.studyDBservice.SaveParticipantState(instanceID, studyKey, actionData.PState)
+				if err != nil {
+					logger.Error.Printf("RunRules: %v", err)
+					return status.Error(codes.Internal, err.Error())
+				}
+
+			}
+			return nil
+		},
+	)
+	if err != nil {
+		logger.Error.Println(err)
+	}
+
+	s.SaveLogEvent(req.Token.InstanceId, req.Token.Id, loggingAPI.LogEventType_LOG, constants.LOG_EVENT_RUN_CUSTOM_RULES, fmt.Sprintf("rules run for study %s: %v", req.StudyKey, req.Rules))
+	resp := api.RuleRunSummary{
+		ParticipantCount:              counters.Participants,
+		ParticipantStateChangePerRule: counters.ParticipantStateChangePerRule,
+		Duration:                      time.Now().Unix() - start,
+	}
+	return &resp, nil
+}
+
 func (s *studyServiceServer) RunRules(ctx context.Context, req *api.StudyRulesReq) (*api.RuleRunSummary, error) {
 	if req == nil || token_checks.IsTokenEmpty(req.Token) || req.StudyKey == "" {
 		return nil, status.Error(codes.InvalidArgument, "missing argument")
@@ -712,8 +858,8 @@ func (s *studyServiceServer) RunRules(ctx context.Context, req *api.StudyRulesRe
 		req.StudyKey,
 		"",
 		func(dbService *studydb.StudyDBService, p types.ParticipantState, instanceID, studyKey string, args ...interface{}) error {
-			if p.StudyStatus == types.PARTICIPANT_STUDY_STATUS_TEMPORARY {
-				// ignore temporary participants
+			if p.StudyStatus != types.PARTICIPANT_STUDY_STATUS_ACTIVE {
+				//CHANGED: ignore all non-active participants
 				return nil
 			}